Avoid downloading avtars everytime on docs:dev
